refactor(updater): separate update actions from modal UI
