+++ docs/frontend-ui-audit-2026-09-07/ComponentDependenciesUpdater.md
@@ -1,0 +1,12 @@
+# ComponentDependenciesUpdater UI audit
+
+| Line                                                              | Element                          | Verdict          | Reason                                                                                                                               | Suggested change |
+| ----------------------------------------------------------------- | -------------------------------- | ---------------- | ------------------------------------------------------------------------------------------------------------------------------------ | ---------------- |
+| `src/scaffold/AppUpdater/index.tsx:57`                            | Install confirmation             | keep with reason | Keeps the existing Modal, Button actions, translations, explicit confirmation, and dimensions; the extraction changes ownership only | None             |
+| `src/scaffold/AppUpdater/DownloadProgress.tsx:38`                 | Download progress notice and orb | keep with reason | Existing progress UI is unchanged; only the progress model import moves into the shared state module                                 | None             |
+| `src/scaffold/NavigationSidebar/blocks/SidebarUpdateButton.tsx:8` | Sidebar updater consumer         | keep with reason | Button and Tooltip composition is unchanged; action and read-hook imports now target their owning modules                            | None             |
+| `src/modules/MainApp/Settings/sections/GeneralSection.tsx:64`     | General settings updater imports | keep with reason | Existing settings controls, copy, and appearance are unchanged                                                                       | None             |
+| `src/engines/ChatPanel/ChatPanelStartPage.tsx:23`                 | ChatPanel update read hook       | keep with reason | Only updater import changed; existing rendering remains owned by ChatPanel                                                           | None             |
+| `src/engines/ChatPanel/hooks/useChatPanelCreationContent.tsx:12`  | ChatPanel update action          | keep with reason | Only updater import changed; no visual structure or accessibility behavior added                                                     | None             |
+
+Verdict totals: **0 fix**, **6 keep with reason**, **0 abstract**.
