+++ src/ActionSystem/actions/updateActions.zod.ts
@@ -12,7 +12,7 @@ import { z } from "zod";
 import { ACTION_ID } from "@src/ActionSystem/actionIds";
 import { defineAppActionRegistration } from "@src/ActionSystem/schema/actionRegistration";
 import { defineZodAction } from "@src/ActionSystem/schema/defineZodAction";
-import { checkForUpdatesManually } from "@src/scaffold/AppUpdater";
+import { checkForUpdatesManually } from "@src/scaffold/AppUpdater/actions";
 
 const emptyParams = z.object({});
