+++ src/engines/ChatPanel/ChatPanelStartPage.lifecycle.test.ts
@@ -23,7 +23,7 @@ const mocks = vi.hoisted(() => ({
   useAvailableAppUpdate: vi.fn(),
 }));
 
-vi.mock("@src/scaffold/AppUpdater", () => ({
+vi.mock("@src/scaffold/AppUpdater/state", () => ({
   useAvailableAppUpdate: mocks.useAvailableAppUpdate,
 }));
 vi.mock("./StartPageQuotaModal", () => ({
