+++ src/engines/ChatPanel/ChatPanelStartPage.tsx
@@ -20,7 +20,7 @@ import {
   Key02Icon,
 } from "@src/icons";
 import { CreatorContentLayout } from "@src/modules/shared/layouts/blocks";
-import { useAvailableAppUpdate } from "@src/scaffold/AppUpdater";
+import { useAvailableAppUpdate } from "@src/scaffold/AppUpdater/state";
 import { creatorComposerPositionAtom } from "@src/store/session/creatorComposerPositionAtom";
 import { creatorLaunchpadActionsVisibleAtom } from "@src/store/session/creatorLaunchpadActionsVisibleAtom";
 import {
